Require all managers to be defined
undefined managers was originally introduced in a refactor that had
code to `initModulesInOrder`; three of the managers could not be
calculated when following that code path. That code was never used and
has since been removed, but the optional manager typing remained.

Static analysis shows no place in the current code where a manager
could be undefined. By marking that in the type system, it simplifies
logic in other locations that would otherwise have to treat each
manager as optional.